+++ packages/models/src/history.rs
@@ -7,5 +7,5 @@ use crate::track::BasicTrack;
 pub struct HistoryItem {
     pub played_at: i64,
     pub track: BasicTrack,
-    pub track_id: i32,
+    pub track_id: u64,
 }
